+++ src/index.jsx
@@ -16,8 +16,8 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import { keyLogin, tryLogin } from './app/actions/auth'
 import App from './app/components/App'
 import Index from './app/components/Index'
-import NewMemberForm from './kansa/components/NewMemberForm'
-import NewMemberIndex from './kansa/components/NewMemberIndex'
+import NewMemberForm from './membership/components/NewMemberForm'
+import NewMemberIndex from './membership/components/NewMemberIndex'
 import PurchaseIndex from './payments/components/PurchaseIndex'
 import PurchaseItem from './payments/components/PurchaseItem'
 import Nominate from './hugo/components/Nominate'
