+++ src/middleware.js
@@ -3,11 +3,11 @@ import { routerMiddleware } from 'react-router-redux'
 
 import hugo from './hugo/middleware'
 import hugoAdmin from './hugo-admin/middleware'
-import kansa from './kansa/middleware'
+import membership from './membership/middleware'
 import payments from './payments/middleware'
 import logger from './app/middleware'
 
 export default (history) => applyMiddleware(
-  hugo, hugoAdmin, kansa, payments, logger,
+  hugo, hugoAdmin, membership, payments, logger,
   routerMiddleware(history)
 );
